requirements.txt
streamlit
numpy
pandas
plotly
altair
regex
python-dotenv
kbcstorage
requests